feat: rename script target to custom
