+++ docs/pages/build.md
@@ -188,17 +188,17 @@ Generates the [React Native Codegen](https://reactnative.dev/docs/the-new-archit
 
 You can ensure your Codegen generated scaffold code is stable through different React Native versions by shipping it with your library. You can find more in the [React Native Official Docs](https://reactnative.dev/docs/the-new-architecture/codegen-cli#including-generated-code-into-libraries).
 
-#### `script`
+#### `custom`
 
-Allows arbitrary scripts to be called during the build process. There is no guarantee when the script is going to be called or finished. This is useful to call code generators during the build process.
+Define a custom build target. This is useful to call code generators during the build process.
 
-##### `run`
+##### `script`
 
-`bob` will run the command passed in the `run` option. The build process will throw and exit if the `script` target is defined without the `run` option.
+`bob` will run the script passed in the `script` option. The build process **will throw and exit** if the `build` target is defined without this option.
 
 ##### `cwd`
 
-You can set the `cwd` (current working directory) option to specify where the command should be called from. This option accepts a path and will default to the path `build` was called from.
+You can set the `cwd` (current working directory) option to specify where the command should be called from. This option accepts **a relative path**, and it will default to the path `build` was called from.
 
 ##### `clean`
