+++ packages/react-native-builder-bob/src/types.ts
@@ -13,7 +13,7 @@ export type Input = {
   report: Report;
 };
 
-export type Target = 'commonjs' | 'module' | 'typescript' | 'codegen' | 'script';
+export type Target = 'commonjs' | 'module' | 'typescript' | 'codegen' | 'custom';
 
 export type Options = {
   source?: string;
